Problem 1: Catch up problem
An arrow has been shoot. A dog saw the arrow and its vision was vertical to the arrow and its moving trail, instantly it ran to catch the arrow. Suppose that the speed of the arrow is v1 and the dog is v2, unchanged, the distance between them when the dog saw the arrow was l, then what's the moving trail of the dog?