Problem 2: Four Dancers in a Square Room
There are four dancers called A, B, C and D, which are now at the edges of same square room clockwise. After started, they dance and move towards each other (A→B, B→C, C→D, D→A) anytime. Suppose that the square room is a unit long and a unit wide, their speeds are same and constant. Please work out each of their moving trails.( By the time the code released this problem had been solved in a formula)